+++ bernard/loader.py
@@ -29,82 +29,12 @@ def build_regex(commands):
     return re.compile("^(" + joined_commands + ")$", re.IGNORECASE)
 
 
-def load_subreddit_rules(subreddit, rules, header, action_buffer_builder,
-                         database):
-    """Create a remover/notifier for each post rule."""
-    post_rules = [
-        rule for rule in rules
-        if rule['kind'] == 'link' or rule['kind'] == 'all'
-    ]
-    our_rules = []
-
-    for i, rule in enumerate(post_rules, 1):
-        note_text = "**{short_name}**\n\n{desc}".format(
-            short_name=rule['short_name'], desc=rule['description'])
-        if header is not None:
-            note_text = header + "\n\n" + note_text
-
-        command = build_regex(
-            ["RULE {n}".format(n=i), "{n}".format(n=i), rule['short_name']])
-        actions = [
-            actors.Notifier(
-                text=note_text, subreddit=subreddit, database=database),
-            actors.ToolboxNoteAdder(
-                text="Post removed (Rule {})".format(i),
-                level="abusewarn",
-                buffer=action_buffer_builder.get(
-                    actors.ToolboxNoteAdderActionBuffer),
-                subreddit=subreddit,
-                database=database)
-        ]
-
-        our_rules.append(
-            actors.Rule(command, [praw.models.Submission], True, True, actions,
-                        'Removed', 'Rule {}'.format(i), database, subreddit))
-
-    return our_rules
-
-
 def validate_actor_config(actor_class, params, target_types):
     """Raise exception if actor configuration is invalid."""
     if not set(target_types).issubset(actor_class.VALID_TARGETS):
         raise RuntimeError("{cls} does not support all of {targets}"
                            .format(cls=actor_class, targets=target_types))
     actor_class.validate_params(params)
-
-
-def load_comment_rules(subreddit, rules, action_buffer_builder, database):
-    """Create a nuker/warner for each comment rule."""
-    header = "Please bear in mind our commenting rules:"
-    comment_rules = [rule for rule in rules if rule['kind'] == 'comment']
-    our_rules = []
-
-    for i, rule in enumerate(comment_rules, 1):
-        note_text = ">**{short_name}**\n\n>{desc}".format(
-            short_name=rule['short_name'], desc=rule['description'])
-        if header is not None:
-            note_text = header + "\n\n" + note_text
-
-        command = build_regex(["nuke {n}".format(n=i), "n {n}".format(n=i)])
-        actions = [
-            actors.Nuker(subreddit=subreddit, database=database),
-            actors.Notifier(
-                text=note_text, subreddit=subreddit, database=database),
-            actors.ToolboxNoteAdder(
-                text="Post removed (Rule {})".format(i),
-                level="abusewarn",
-                buffer=action_buffer_builder.get(
-                    actors.ToolboxNoteAdderActionBuffer),
-                subreddit=subreddit,
-                database=database)
-        ]
-
-        our_rules.append(
-            actors.Rule(command, [praw.models.Comment], True, False, actions,
-                        'Nuked', 'Comment Rule {}'.format(i), database,
-                        subreddit))
-
-    return our_rules
 
 
 class ActionBufferBuilder:
@@ -172,8 +102,7 @@ class YAMLLoader:
         if hasattr(actor_class, 'ACTION_BUFFER'):
             params['buffer'] = action_buffer_builder.get(
                 actor_class.ACTION_BUFFER)
-        return actor_class(
-            database=self.database, subreddit=subreddit, **params)
+        return actor_class(subreddit=subreddit, **params)
 
     def parse_subreddit_config(self, subreddit_config):
         """Parse subreddit configuration and return a Browser."""
@@ -184,21 +113,6 @@ class YAMLLoader:
                                    action_buffer_builder)
             for rule_config in subreddit_config['rules']
         ]
-        header = subreddit_config.get('header')
-        sub_rules = subreddit.rules()['rules']
-        # TODO
-
-        # Add remover/notifier for subreddit rules and, if desired,
-        # nuker/notifier for comment rules.
-        if subreddit_config.get('default_post_actions'):
-            browser_rules.extend(
-                load_subreddit_rules(subreddit, sub_rules, header,
-                                     action_buffer_builder, self.database))
-
-        if subreddit_config.get('default_comment_actions'):
-            browser_rules.extend(
-                load_comment_rules(subreddit, sub_rules, action_buffer_builder,
-                                   self.database))
 
         helpers.update_sr_tables(self.database.cursor(), subreddit)
         self.database.commit()