+++ CHANGES.md
@@ -5,3 +5,5 @@
 * Added option to leave removed post unlocked
 * Renamed 'wikiwatch' to 'userwatch' for configuration file.
 * Added 'domainwatch' to add post domains to AutoMod configuration.
+* Removed legacy default configuration loader
+* Removed notifications and removals from database